Move statics for `GetSpec` to `SharedApiTypes`
